fix: update import outside v2
