+++ src/components/[guild]/JoinModal/components/WalletAuthButton.tsx
@@ -1,4 +1,4 @@
-import { walletSelectorModalAtom } from "@/components/Providers/Providers"
+import { walletSelectorModalAtom } from "@/components/Providers/atoms"
 import { useWeb3ConnectionManager } from "@/components/Web3ConnectionManager/hooks/useWeb3ConnectionManager"
 import { SignIn, Wallet } from "@phosphor-icons/react"
 import { useSetAtom } from "jotai"
