+++ src/components/[guild]/Requirements/components/GuildCheckout/components/buttons/ConnectWalletButton.tsx
@@ -1,4 +1,4 @@
-import { walletSelectorModalAtom } from "@/components/Providers/Providers"
+import { walletSelectorModalAtom } from "@/components/Providers/atoms"
 import { useWeb3ConnectionManager } from "@/components/Web3ConnectionManager/hooks/useWeb3ConnectionManager"
 import { Collapse, Tooltip } from "@chakra-ui/react"
 import Button from "components/common/Button"
