+++ src/components/common/Layout/components/Account/Account.tsx
@@ -1,4 +1,4 @@
-import { walletSelectorModalAtom } from "@/components/Providers/Providers"
+import { walletSelectorModalAtom } from "@/components/Providers/atoms"
 import { useWeb3ConnectionManager } from "@/components/Web3ConnectionManager/hooks/useWeb3ConnectionManager"
 import { ButtonGroup, Divider } from "@chakra-ui/react"
 import { SignIn } from "@phosphor-icons/react"
