+++ src/v2/components/Providers/types.ts
@@ -1,0 +1,4 @@
+export type AddressLinkParams = {
+  userId?: number
+  address?: `0x${string}`
+}
