+++ src/v2/components/Web3ConnectionManager/WalletSelectorModal/hooks/useLinkAddress.ts
@@ -1,8 +1,6 @@
+import { addressLinkParamsAtom } from "@/components/Providers/atoms"
 import { usePostHogContext } from "@/components/Providers/PostHogProvider"
-import {
-  AddressLinkParams,
-  addressLinkParamsAtom,
-} from "@/components/Providers/Providers"
+import { AddressLinkParams } from "@/components/Providers/types"
 import { useWeb3ConnectionManager } from "@/components/Web3ConnectionManager/hooks/useWeb3ConnectionManager"
 import { useUserPublic } from "@/hooks/useUserPublic"
 import useSubmit from "hooks/useSubmit"
